skyeng/requirements.txt
dagster>=1.1.20
dagit>=1.1.20
dagster-postgres>=0.17.13
psycopg2-binary>=2.9.5
clickhouse-sqlalchemy>=0.2.3
pytest>=7.2.0
cryptography>=39.0.0
pandas>=1.5.3
clickhouse-connect>=0.5.7